+++ Content/MorphBallAddons/BoostBall.cs
@@ -15,7 +15,7 @@ namespace MetroidMod.Content.MorphBallAddons
 
 		public override bool CanGenerateOnChozoStatue(int x, int y) => true;
 
-		public override int GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 7;
+		public override double GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 7;
 
 		public override void SetStaticDefaults()
 		{
