+++ Content/MorphBallAddons/FireBomb.cs
@@ -16,7 +16,7 @@ namespace MetroidMod.Content.MorphBallAddons
 
 		public override bool CanGenerateOnChozoStatue(int x, int y) => WorldGen.drunkWorldGen || y >= Main.UnderworldLayer;
 
-		public override int GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 30;
+		public override double GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 30;
 
 		public override void SetStaticDefaults()
 		{
