+++ Content/MorphBallAddons/PowerBomb.cs
@@ -26,7 +26,7 @@ namespace MetroidMod.Content.MorphBallAddons
 
 		public override bool CanGenerateOnChozoStatue(int x, int y) => WorldGen.drunkWorldGen;
 
-		public override int GenerationChance(int x, int y) => 20;
+		public override double GenerationChance(int x, int y) => 20;
 
 		public override void SetStaticDefaults()
 		{
