+++ Content/SuitAddons/PowerGrip.cs
@@ -16,7 +16,7 @@ namespace MetroidMod.Content.SuitAddons
 
 		public override bool CanGenerateOnChozoStatue(int x, int y) => true;
 
-		public override int GenerationChance(int x, int y) => 20;
+		public override double GenerationChance(int x, int y) => 20;
 
 		public override void SetStaticDefaults()
 		{
