+++ Content/SuitAddons/ReserveTank.cs
@@ -17,7 +17,7 @@ namespace MetroidMod.Content.SuitAddons
 
 		public override bool CanGenerateOnChozoStatue(int x, int y) => true;
 
-		public override int GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 10;
+		public override double GenerationChance(int x, int y) => WorldGen.drunkWorldGen ? 20 : 10;
 
 		public override void SetStaticDefaults()
 		{
